use new list for listing right holders
