+++ src/features/amUI/users/UsersList.module.css
@@ -1,3 +1,7 @@
+.usersList {
+  padding: 1rem 0;
+}
+
 .listItemContent {
   display: flex;
   justify-content: flex-start;
@@ -13,4 +17,10 @@
   display: flex;
   flex-wrap: wrap;
   gap: 0.5rem;
+}
+
+.pagination {
+  display: flex;
+  justify-content: center;
+  padding-top: 1rem;
 }
